Move tests as per Code Review
