--- tests/derive_zero.rs
@@ -1,28 +1,0 @@
-//! This module exercises the `ff_derive` procedural macros, specifically to ensure
-//! zeroize works
-
-#[macro_use]
-extern crate ff;
-
-/// The BLS12-381 scalar field.
-#[derive(PrimeField)]
-#[PrimeFieldModulus = "52435875175126190479447740508185965837690552500527637822603658699938581184513"]
-#[PrimeFieldGenerator = "7"]
-#[PrimeFieldReprEndianness = "little"]
-struct Bls381K12Scalar([u64; 4]);
-
-#[test]
-fn zeroize() {
-    use ff::{Field, PrimeField};
-    use rand::rngs::OsRng;
-    use zeroize::Zeroize;
-
-    let mut f = Bls381K12Scalar::random(OsRng);
-    let mut r = f.to_repr();
-
-    f.zeroize();
-    assert_eq!(f, Bls381K12Scalar::ZERO);
-
-    r.zeroize();
-    assert_eq!(r, Default::default());
-}
